requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
docx==0.2.4
et-xmlfile==1.0.1
Flask==1.0.3
fpdf==1.7.2
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.18
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kiwisolver==1.1.0
lxml==4.3.4
MarkupSafe==1.1.1
matplotlib==3.1.0
more-itertools==7.0.0
networkx==2.3
numpy==1.16.4
oauthlib==3.0.1
opencv-python==4.1.0.25
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pdfminer3k==1.3.1
Pillow==6.2.0
pkg-resources==0.0.0
pluggy==0.12.0
ply==3.11
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pymssql==2.1.4
PyMySQL==0.9.3
pyparsing==2.4.0
PyPDF2==1.26.0
pytest==4.6.3
python-dateutil==2.8.0
python-docx==0.8.10
python-http-client==3.1.0
pytz==2019.1
PyWavelets==1.0.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-image==0.15.0
scipy==1.3.0
sendgrid==6.0.5
six==1.12.0
slate3k==0.5.3
soupsieve==1.9.1
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
xlrd==1.2.0
XlsxWriter==1.1.8
zipp==0.5.1
